requirements.txt
streamlit
openai
pytesseract
pillow
geopy
pdfplumber
fitz  # or pymupdf
python-dotenv
